Describe whether variables are completely missing for each cohort

When performing a Study Level Meta Analysis of coefficients, it is usually necessary that all cohorts have some data on all variables in the model. Manually identifying which cohorts have some data on required variables is time-consuming and at risk of error. This function automatically generates a look-up table with this information.

dh.anyData(df = NULL, vars = NULL, conns = NULL, checks = TRUE)

Arguments

df

Character specifying a server-side data frame.

vars

Character vector of columns within `df`.

conns

DataSHIELD connections object.

checks

Logical; if TRUE checks are performed prior to running the function. Default is TRUE.

Value

Client-side tibble containing columns for each variable and rows for each cohort indicating if the variable is not completely missing.

See also