This was the original version of dh.makeStrata. It has now been deprecated.

dh.makeStrata contains a number of improvements: it is computationally more efficient, it is more clearly coded, and it has been more thoroughly tested.

dh.makeOutcome(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  bands = NULL,
  mult_action = NULL,
  mult_vals = NULL,
  df_name = NULL,
  id_var = "child_id",
  band_action = NULL,
  keep_original = NULL,
  conns = NULL
)

Arguments

df

See argument `df` in `dh.makeStrata`.

outcome

Renamed to `var_to_subset` in `dh.makeStrata`

age_var

See argument `age_var` in `dh.makeStrata`.

bands

See argument `df` in `dh.makeStrata`

mult_action

See argument `mult_action` in `dh.makeStrata`

mult_vals

See argument `mult_vals` in `dh.makeStrata`

df_name

Renamed to `new_obj` in `dh.makeStrata`

id_var

See argument `id_var` in `dh.makeStrata`

band_action

See argument `band_action` in `dh.makeStrata`

keep_original

Argument no longer in use in `dh.makeStrata`

conns

See argument `conns` in `dh.makeStrata`