This was the original version of dh.meanByGroup. It has now been deprecated.

Whilst I always use this with 'age' as a grouping variable, it is a more general function than that. I have thus renamed it to reflect that any continuous grouping variable can be used.

dh.meanByAge(
  df = NULL,
  outcome = NULL,
  age_var = NULL,
  intervals = NULL,
  conns = NULL,
  checks = FALSE
)

Arguments

df

Character specifying a server-side data frame.

outcome

outcome variable in long format

age_var

age in years

intervals

table defining our age bands

conns

DataSHIELD connections object.

checks

Logical; if TRUE checks are performed prior to running the function. Default is TRUE.

Value

Mean values for each unit of your age variable are returned