Removes columns from a serverside data frame

This function allowed you to subset a data frame by column names. It was deprecated because you can now use dsTidyverseClient::ds.select() which is much quicker and has greater flexibility.

Usage

dh.dropCols(
  df = NULL,
  vars = NULL,
  new_obj = NULL,
  type = NULL,
  conns = NULL,
  checks = TRUE,
  new_df_name = NULL
)

Arguments

df

Character specifying a server-side data frame.

vars

Character vector specifying columns within df to be removed or kept.

new_obj

Optionally, character specifying name for new server-side data frame. Default is to return original data frame with columns removed.

type

Character specifying how to treat vars. If "remove" these variables are removed from the data frame, if "keep" these variables are kept in the data frame and all others are removed.

conns

DataSHIELD connections object.

checks

Logical; if TRUE checks are performed prior to running the function. Default is TRUE.

new_df_name

Retired argument name. Please use `new_obj' instead.

Value

Server-side data frame the specified subset of columns.