Extracts the coefficients and confidence intervals from a ds.glm or ds.glmSLMA model.

Extracts the coefficients and confidence intervals from a ds.glm or ds.glmSLMA model.

dh.lmTab(
  model = NULL,
  type = NULL,
  coh_names = NULL,
  direction = c("long", "wide"),
  ci_format = NULL,
  family = "gaussian",
  round_digits = 2,
  exp = FALSE
)

Arguments

model
saved output from either ds.glm, ds.glmSLMA or ds.lmerSLMA

type
either "ipd", "slma" or "lmer"

coh_names
a vector of cohort names. Note this needs to be in the same order as the cohorts provided to the model.

direction
either "long" or "wide"

ci_format
format for the confidence intervals when direction == "wide". "separate" outputs separate columns with upper and lower CIs. "paste" adds these in brackets to the coefficient.'

family
where type is ipd or slma, specify the family used in the model

round_digits
Number of decimal places to use in table. Default is 2.

exp
specify whether you want odds ratios to be exponentiated

Value

A tibble containing the estimate with lower and upper confidence intervals