Make formulae for fitting multiple fractional polynomial models

To identify the combination of fractional polynomials which give the best fit often we will fit models with lots of different combinations. This function builds formulae for that purpose.

dh.makeLmerForm(
  outcome = NULL,
  idvar = NULL,
  agevars = NULL,
  random = NULL,
  fixed = NULL,
  age_interactions = NULL
)

Arguments

outcome
outcome for the models

idvar
unique identifier for subject

agevars
vector of names of age polynomials in dataset that you will use in your models

random
either "intercept" or "slope" to specify random effects

fixed
optional vector of fixed effects

age_interactions
if TRUE also create interaction terms between age terms and fixed effects

Value

a tibble containing the created formulae