Trims predicted values based on min and max values provided. Useful where you want to ensure that predicted values are limited at the range of observed values.

Trims predicted values based on min and max values provided. Useful where you want to ensure that predicted values are limited at the range of observed values.

dh.trimPredData(
  pred = NULL,
  coh_names = NULL,
  age = "age",
  min = NULL,
  max = NULL
)

Arguments

pred
a tibble of predicted values

coh_names
a vector of cohort names corresponding to the cohort names in pred

age
name of the variable corresponding to age in the predicted values

min
a vector of minimum values of age, length equal to number of cohorts

max
a vector of maximum values of age, length equal to number of cohorts

Value

The original tibble of predicted values, trimmed by the min and max ages